+++ src/components/Home.tsx
@@ -1,21 +1,17 @@
-import { useContext } from 'react';
+// import { useContext } from 'react';
 import Nav from './Nav';
-import UserContext from '../contexts/UserContext';
+// import UserContext from '../contexts/UserContext';
+import getToken from '../utils/getToken';
 
 export default function Home() {
-  const { user } = useContext(UserContext);
-  const token = localStorage.getItem('user');
-  let tokenObj;
-  if (token) {
-    tokenObj = JSON.parse(token);
-  }
-  const tokenStr = tokenObj.token;
-  console.log(user);
+  // const { user } = useContext(UserContext);
+  const tokenObj = getToken();
+
   return (
     <div>
      <Nav />
      Home
-     <div> { tokenStr
+     <div> { tokenObj
        ? (`${tokenObj.email} is logged in`) : 'You are not logged in.'
      } </div>
     </div>
