+++ src/components/LogIn.tsx
@@ -1,25 +1,17 @@
-import { useContext } from 'react';
 import Nav from './Nav';
-import UserContext from '../contexts/UserContext';
-// import { login } from '../utils/login';
 import LoginForm from '../utils/loginForm';
+import getToken from '../utils/getToken';
 
 export default function LogIn() {
-  const { user } = useContext(UserContext);
-  const token = localStorage.getItem('user');
-  let tokenObj;
-  if (token) {
-    tokenObj = JSON.parse(token);
-  }
-  const tokenStr = tokenObj.token;
-  console.log(user);
+  const tokenObj = getToken();
   return (
     <div>
      <Nav />
      Home
-     <div> { tokenStr
-       ? (`${tokenObj.email} is logged in`) : 'You are not logged in.'
-     } <LoginForm/> </div>
+     <div> { tokenObj
+       ? (`${tokenObj.email} is logged in`) : (
+        <LoginForm />)}
+      </div>
     </div>
   );
 }
