+++ src/components/SignUp.tsx
@@ -1,15 +1,18 @@
-import { useContext } from 'react';
+// import { useContext } from 'react';
 import Nav from './Nav';
-import UserContext from '../contexts/UserContext';
+// import UserContext from '../contexts/UserContext';
+import getToken from '../utils/getToken';
 
 export default function SignUp() {
-  const { user } = useContext(UserContext);
+  // const { user } = useContext(UserContext);
+  const tokenObj = getToken();
+
   return (
     <div>
      <Nav />
      Sign Up
-     <div> { user
-       ? (`${JSON.stringify(user, null, 2)} is logged in`) : 'You are not logged in.'
+     <div> { tokenObj
+       ? (`${tokenObj.email} is logged in`) : 'You are not logged in.'
      } </div>
     </div>
   );
