+++ src/utils/logOut.tsx
@@ -1,0 +1,3 @@
+export default function logOut() {
+  localStorage.setItem('user', '');
+}
